// src/types/constants.ts
// Enums matching backend
export enum BloodType {
  A_POSITIVE = 'A+',
  A_NEGATIVE = 'A-',
  B_POSITIVE = 'B+',
  B_NEGATIVE = 'B-',
  AB_POSITIVE = 'AB+',
  AB_NEGATIVE = 'AB-',
  O_POSITIVE = 'O+',
  O_NEGATIVE = 'O-'
}

export enum BloodComponent {
  WHOLE_BLOOD = 'whole_blood',
  RED_CELLS = 'red_cells',
  PLATELETS = 'platelets',
  PLASMA = 'plasma'
}